debugging mesosim vaporizer
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5322 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSNet.cpp
@@ -181,7 +181,7 @@ MSNet::closeBuilding(MSEdgeControl *edges, MSJunctionControl *junctions,
 
     // we may add it before the network is loaded
     if (myEdges!=0) {
-        myEdges->insertMeanData(myMeanData.size());
+        myEdges->insertMeanData((unsigned int) myMeanData.size());
     }
 #ifdef HAVE_MESOSIM
     if (MSGlobals::gUseMesoSim) {
